Rename folder Event to Events
Follow the pattern of other apps and use the folder name in plural
