+++ lib/Events/SignedEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OCA\Libresign\Event;
+namespace OCA\Libresign\Events;
 
 class SignedEvent extends ASignedEvent {
 }
